extract export features for use in init ⌘
